requirements.txt
streamlit
pandas
PyPDF2
pdfplumber
# Optional (for Gemini Q&A)
google-genai